Crate risc0_circuit_rv32im

The RISC Zero zkVM circuit

Crate Feature Flags

The following crate feature flags are available.

Feature	Target(s)	Implies	Description

cuda		prove, std	Turns on CUDA GPU acceleration for the prover. Requires CUDA toolkit to be installed.

metal	macos	prove, std	Deprecated - Metal GPU acceleration for the prover is enabled by default on Apple Silicon.

prove	all except rv32im	std	Enables the prover, incompatible within the zkvm guest.

std	all		Support for the Rust stdlib.

Modules

control_id

layout

poly_ext

trace

Structs

CircuitImpl

Constants

CIRCUIT

GLOBAL_MIX

GLOBAL_OUT

REGISTER_GROUP_ACCUM

REGISTER_GROUP_CODE

REGISTER_GROUP_CTRL

REGISTER_GROUP_DATA

Functions

control_id

Fetch a control ID with the given hash, by name, and cycle limit as a power of two (po2) from the precomputed table. If the hash function is not precomputed, or the po2 is out of range, this function will return None.

control_ids

Fetch all precomputed control IDs using the given hash, by name, and up to the cycle limit as a power of two (po2). If po2 is larger than the max supported, only supported po2s will be returned.